add colors to project, clean main activity and restructure viewModel/screen
