+++ app/src/main/java/com/example/screp/ui/theme/Color.kt
@@ -2,7 +2,9 @@ package com.example.screp.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val Purple200 = Color(0xFFBB86FC)
-val Purple500 = Color(0xFF6200EE)
-val Purple700 = Color(0xFF3700B3)
-val Teal200 = Color(0xFF03DAC5)
+val Primary = Color(0xFF2C3333)
+val Secondary = Color(0xFF395B64)
+val Background = Color(0xFFE7F6F2)
+val NavBarIcon = Color(0xFFEBEBEB)
+val NavBarChosenIcon = Color(0xFFE3C770)
+val LinearBackgroundLight = Color(0xFF6CB5A4)
