+++ app/src/main/java/com/example/screp/ui/theme/Theme.kt
@@ -5,17 +5,21 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 private val DarkColorPalette = darkColors(
-    primary = Purple200,
-    primaryVariant = Purple700,
-    secondary = Teal200
+    primary = Primary,
+    primaryVariant = LinearBackgroundLight,
+    secondary = Secondary,
+    background = Background
 )
 
 private val LightColorPalette = lightColors(
-    primary = Purple500,
-    primaryVariant = Purple700,
-    secondary = Teal200
+    primary = Primary,
+    primaryVariant = LinearBackgroundLight,
+    secondary = Secondary,
+    background = Background
 
     /* Other default colors to override
     background = Color.White,
@@ -33,6 +37,14 @@ fun ScrepTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable 
         DarkColorPalette
     } else {
         LightColorPalette
+    }
+
+    val systemUiController = rememberSystemUiController()
+    SideEffect {
+        systemUiController.setSystemBarsColor(
+            color = Primary,
+            darkIcons = false
+        )
     }
 
     MaterialTheme(
